functions/requirements.txt
firebase-functions>=0.1.0
firebase-admin>=6.0.0
google-cloud-firestore>=2.13.0
google-cloud-storage>=2.10.0
google-generativeai>=0.7.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
flask>=2.3.0
